refactor: split Overlay's useOverlay into focused modules
useOverlay.tsx grew to 858 lines mixing pure positioning math with React
state, focus management, and 5+ event-listener effects. Split into modules
so each concern is independently readable and testable:

- positionMath.ts (301 lines): pure math — calcDropdown{Vertical,Horizontal},
  calcModalPos, adjustForAncestor, computePosition, processVisibilityEvent.
  No React, no DOM mutations. Type definitions for Align/AlignX/AlignY/
  OverlayType/OverlayPosition exported alongside.

- useEscapeKey.ts (22 lines): ESC key handler.

- useHoverListeners.ts (109 lines): mouseenter/mouseleave on trigger +
  content with hoverDelay timeout, including the bridge logic that prevents
  flicker when moving from trigger to content.

- useScrollReposition.ts (118 lines): scroll/resize listeners on window and
  optional parentContainer, plus the refcounted modal body-overflow lock.

- useOverlay.tsx (448 lines, was 858): now orchestrates state, positioning
  callbacks, focus management, and the lifecycle effect, then composes the
  extracted hooks.

Public API unchanged. All 74 Overlay tests + 268 elements tests + 2599 total
tests pass. 0 type errors, 0 lint errors, 15 packages build.
